Core/Misc: Fixed nopch MSVC build
